Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -13,7 +13,7 @@ int main() {
         printf("%d",a*b);
     }
     else if (operators == '/') {
-        printf("%d",(float)a/b);
+        printf("%d",a/b);
     }
     else {
         printf("Yoyo");
